refactor: fix imports from JobExecutionHelper to JobExecutionWaitUtils (#750)
- Replace incorrect imports from JobExecutionHelper with JobExecutionWaitUtils
- Reduce default checkInterval to 250ms to increase test speed
