+++ engine-cdi/src/test/java/org/operaton/bpm/engine/cdi/test/impl/event/EventNotificationTest.java
@@ -26,7 +26,7 @@ import org.operaton.bpm.engine.test.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.operaton.bpm.engine.test.util.JobExecutorHelper;
+import org.operaton.bpm.engine.test.util.JobExecutorWaitUtils;
 
 import java.util.List;
 import java.util.Set;
@@ -133,7 +133,7 @@ public class EventNotificationTest extends CdiProcessEngineTestCase {
 
     assertThat(listenerBean.getEventsReceived()).isEmpty();
     runtimeService.startProcessInstanceByKey("process1");
-    JobExecutorHelper.waitForJobExecutorToProcessAllJobs(processEngineConfiguration, TimeUnit.SECONDS.toMillis(5L), 500L);
+    JobExecutorWaitUtils.waitForJobExecutorToProcessAllJobs(processEngineConfiguration, TimeUnit.SECONDS.toMillis(5L), 500L);
 
     Task task = taskService.createTaskQuery().singleResult();
     assertThat(task.getName()).isEqualTo("User Task");
